fixed bug in vms-ovirt/clean task where it would return a 404.
Slight regression due to side effect of update in ovirtsdk4.

In ovirtsdk4 4.1.0 this code was changed in service.py:

  @@ -94,7 +94,7 @@
		   xmlreader.close()
	       if buf is not None:
		   buf.close()
  -        if fault is not None:
  +        if fault is not None or response.code >= 400:
	       Service._raise_error(response, fault)

This causes the ovirt utility in ansible to fail as part of it's normal remove
operations, as it performs these operations:

  * GET :vm_id:
  * DELETE :vm_id:
  * GET :vm_id:

The final GET worked in ovirtsdk4 4.0.4, as the 404 was silently ignored. Now
in 4.1.0 it raises an exception.